// src/components/SetupIncidentResponseFlow/enums.ts
// @generated by drool.exe extraction
// Generated from src/components/SetupIncidentResponseFlow/enums; source module was missing in Bun sourcemap recovery.
export enum WizardStep {
  Applying = "applying", // high-inferred; compactCamel; bundle hits=5
  Channel = "channel", // medium-cluster; compactCamel; bundle hits=58
  Computer = "computer", // high-inferred; compactCamel; bundle hits=49
  Confirm = "confirm", // medium-cluster; compactCamel; bundle hits=30
  Failed = "failed", // medium-cluster; compactCamel; bundle hits=73
  LoadFailed = "loadFailed", // high-inferred; compactCamel; bundle hits=5
  Loading = "loading", // medium-cluster; compactCamel; bundle hits=37
  Prompt = "prompt", // medium-cluster; compactCamel; bundle hits=61
  Success = "success", // medium-cluster; compactCamel; bundle hits=81
}